86d05e3c0 - Move the landlord task creation into the slice
